Generate and use a static screenshot of the gif
Hooks in to generation of image metadata and manually generate a
static copy of the gif (by flipping the image twice, which destroys
all but the first animation frame), then insert this new size
into the metadata array.

When returning the JSON response, make sure our image is included,
by checking a special flag that asks for it to be included.

In the gif view, use the static URL first, and then on mouseover, swap
it out for the animated view (going back to static on mouseout).