Resize and crop images so they fit really nicely.
On hover, it will either expose the bottom part of the cropped image,
if there isn't much to show, or it will shrink the image down
proportionally within its box to show it all, if the image is a lot
taller than we're showing.

Downside: images that are really far from ratios of 1/2, 1/1, or 2/1
will be cropped a good bit.

Upside: they fix together **really** nicely now.

fixes #4